update binance TD with query order,account trade list

+++ longfist/longfist/LFDataStruct.h
@@ -248,11 +248,11 @@ struct LFQryPositionField
 struct LFRspPositionField
 {
 	char_31              	InstrumentID;          //合约代码
-	int                  	YdPosition;            //上日持仓
-	int                  	Position;              //总持仓
+    uint64_t                  	YdPosition;            //上日持仓
+	uint64_t                  	Position;              //总持仓
 	char_11              	BrokerID;              //经纪公司代码
 	char_19              	InvestorID;            //投资者代码
-	double               	PositionCost;          //持仓成本
+	int64_t               	PositionCost;          //持仓成本
 	LfHedgeFlagType      	HedgeFlag;             //投机套保标志
 	LfPosiDirectionType  	PosiDirection;         //持仓多空方向
 };
